Several MP fixes, resolve a TODO, cleanup

+++ Projectiles/Minions/GoblinTechnomancer/GoblinTechnomancer.cs
@@ -403,13 +403,15 @@ namespace AmuletOfManyMinions.Projectiles.Minions.GoblinTechnomancer
 			Projectile.rotation = Projectile.velocity.X * 0.025f;
 		}
 
-		public override void CheckActive()
+		public override bool CheckActive()
 		{
-			base.CheckActive();
-			if(player.ownedProjectileCounts[CounterType] == 0 && animationFrame > 2)
+			if (base.CheckActive() && player.ownedProjectileCounts[CounterType] == 0 && animationFrame > 2)
 			{
 				Projectile.Kill();
+				return false;
 			}
+
+			return true;
 		}
 	}
 }
